refactor: Use related news API
