+++ .env.development
@@ -1,7 +1,7 @@
 NEXT_PUBLIC_ALGOLIA_APP_ID=XQEP3AD9ZT
 NEXT_PUBLIC_ALGOLIA_API_KEY=118b51a67eefbe4d3c445d1e1090afac
 NEXT_PUBLIC_SHOPIFY_DOMAIN=ifixit-test.myshopify.com
-NEXT_PUBLIC_SHOPIFY_STOREFRONT_ACCESS_TOKEN=bc6c5d0972b3ad5d565c99768e6065bf
+NEXT_PUBLIC_SHOPIFY_STOREFRONT_ACCESS_TOKEN=d08db995952e88a37804b54a284e5130
 NEXT_PUBLIC_IFIXIT_API_ORIGIN=https://www.cominor.com
 NEXT_PUBLIC_WP_ORIGIN=https://valkyrie.ifixit.com
 NEXT_PUBLIC_WP_BASIC_AUTH_USER=ifixit
